change: extract data loader from engine to split complexity
