+++ .gitignore
@@ -44,3 +44,5 @@ client/.npm
 /server/server-artifact/data/
 /client/nbproject/private/
 /server/server-client-dummie/target/
+/server/server-data-loader/target/
+/server/server-data-loader/data/
