+++ server/server-data-loader/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.thegame.server</groupId>
+        <artifactId>server</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>server-data-loader</artifactId>
+    <packaging>jar</packaging>
+    <name>TheGame Server Data Loader</name>
+	
+	<dependencies>
+		<dependency>
+			<groupId>com.thegame.server</groupId>
+			<artifactId>server-engine</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+	</dependencies>
+</project>
