+++ server/server-data-loader/src/main/java/com/thegame/server/data/loader/beans/ItemInstanceImport.java
@@ -1,4 +1,4 @@
-package com.thegame.server.engine.intern.load.beans;
+package com.thegame.server.data.loader.beans;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
