+++ server/server-data-loader/src/test/resources/load/area/area2.json
@@ -1,0 +1,10 @@
+{
+	"id": "beta-room-002",
+	"title": "roomBeta001Title",
+	"description": "roomBeta001Description",
+	"exits": {
+		"north": "beta-room-002",
+		"south": "beta-room-002"
+	},
+    "items": []	
+}
