+++ server/server-data-loader/src/test/resources/load/item/item.json
@@ -1,0 +1,11 @@
+{
+    "id": "chicken-001",
+    "description": "itemChickenDescription",
+    "alive": true,
+    "gender": "male",
+    "singular": "itemChickenSingular",
+    "plural": "itemChickenPlural",
+    "chatters": [
+        "cooooo!"
+    ]
+}
