+++ server/server-engine/data/thegame.properties
@@ -1,5 +1,5 @@
 #HSQL Database Engine 1.8.0.10
-#Sun May 14 23:55:21 CEST 2017
+#Mon May 15 00:03:40 CEST 2017
 hsqldb.script_format=0
 runtime.gc_interval=0
 sql.enforce_strict_size=false
