--- server/server-engine/src/main/java/com/thegame/server/engine/intern/load/beans/AreaImport.java
@@ -1,31 +1,0 @@
-package com.thegame.server.engine.intern.load.beans;
-
-import com.thegame.server.engine.messages.IsMessageBean;
-import java.util.List;
-import java.util.Map;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.Singular;
-
-/**
- * @author afarre
- */
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AreaImport implements IsMessageBean{
-
-	@NonNull
-	private String id;
-	private String title;
-	private String description;
-	
-	@Singular
-	private Map<String,String> exits;
-	@Singular
-	private List<ItemInstanceImport> items;
-}
